# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Marketplace API",
    "description": "API for managing products and orders in an online marketplace",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://electro-mart-37rz.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current authenticated user",
        "description": "",
        "responses": {
          "200": {
            "description": "Current user returned successfully"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
...
  }
}